IFS-2016-PF-SpendProfile-Action
Fixed Data Layer junits.

+++ ifs-data-layer/ifs-data-service/src/test/java/org/innovateuk/ifs/competition/controller/CompetitionPostSubmissionControllerTest.java
@@ -100,7 +100,7 @@ public class CompetitionPostSubmissionControllerTest extends BaseControllerMockM
     @Test
     public void countPendingSpendProfiles() throws Exception {
         final Long competitionId = 1L;
-        final Integer pendingSpendProfileCount = 3;
+        final Long pendingSpendProfileCount = 3L;
 
         when(competitionServiceMock.countPendingSpendProfiles(competitionId)).thenReturn(serviceSuccess(pendingSpendProfileCount));
