+++ ifs-data-layer/ifs-data-service/src/test/java/org/innovateuk/ifs/competition/security/CompetitionServiceSecurityTest.java
@@ -375,7 +375,7 @@ public class CompetitionServiceSecurityTest extends BaseServiceSecurityTest<Comp
         }
 
         @Override
-        public ServiceResult<Integer> countPendingSpendProfiles(Long competitionId) {
+        public ServiceResult<Long> countPendingSpendProfiles(Long competitionId) {
             return null;
         }
     }
